Moved RecordResults methods to record_results.go
